drop ik solver interface and re-structure utils files
