+++ isaaclab_arena_curobo/ik_reachability_validator.py
@@ -22,11 +22,8 @@ from isaaclab_arena.relations.placement_validators import PlacementValidator
 from isaaclab_arena.relations.relations import get_anchor_objects
 from isaaclab_arena.utils.yaw import rotate_quat_by_yaw, yaw_from_quat_xyzw
 from isaaclab_arena_curobo.ik_solver import CuroboIKSolver
-from isaaclab_arena_curobo.ik_solver_utils import (
-    get_aabb_collision_cuboid_for_object,
-    solve_ik_feasibility,
-    top_down_grasp_pose_from_world_poses,
-)
+from isaaclab_arena_curobo.utils.frame_utils import top_down_grasp_pose_from_world_poses
+from isaaclab_arena_curobo.utils.ik_solver_utils import get_aabb_collision_cuboid_for_object, solve_ik_feasibility
 
 if TYPE_CHECKING:
     from collections.abc import Callable
