+++ isaaclab_arena_curobo/ik_solver.py
@@ -18,9 +18,9 @@ from curobo.wrap.reacher.ik_solver import IKSolver, IKSolverConfig
 
 from isaaclab_arena.utils.device import resolve_cuda_device
 from isaaclab_arena_curobo.curobo_embodiment_cfg import CuroboEmbodimentCfg
-from isaaclab_arena_curobo.curobo_frame_utils import load_patched_robot_yaml
 from isaaclab_arena_curobo.embodiment_curobo_registry import get_curobo_cfg_for
-from isaaclab_arena_curobo.ik_solver_utils import AABBCollisionCuboid, world_config_from_cuboids
+from isaaclab_arena_curobo.utils.ik_solver_utils import AABBCollisionCuboid, world_config_from_cuboids
+from isaaclab_arena_curobo.utils.robot_cfg_utils import load_patched_robot_yaml
 
 
 class CuroboIKSolver:
