+++ isaaclab_arena_curobo/tests/test_curobo_deps.py
@@ -29,9 +29,10 @@ CUROBO_DEPENDENCY_MODULES = [
 
 # The cuRobo-gated package modules, which must import without a running SimulationApp.
 ARENA_CUROBO_MODULES = [
-    "isaaclab_arena_curobo.curobo_frame_utils",
-    "isaaclab_arena_curobo.ik_solver_utils",
-    "isaaclab_arena_curobo.curobo_planner_utils",
+    "isaaclab_arena_curobo.utils.frame_utils",
+    "isaaclab_arena_curobo.utils.ik_solver_utils",
+    "isaaclab_arena_curobo.utils.planner_utils",
+    "isaaclab_arena_curobo.utils.robot_cfg_utils",
     "isaaclab_arena_curobo.placement_pool_ik_validation",
     "isaaclab_arena_curobo.ik_reachability_validator",
 ]
