+++ isaaclab_arena_curobo/tests/test_ik_reachability.py
@@ -153,7 +153,7 @@ def _run_sync_pose_check(args_cli) -> bool:
     # Offset + 90 deg yaw on the robot base so the world->base transform is non-trivial: a bug that
     # left poses in world frame (or used the wrong quaternion order) would then diverge from expected.
     from isaaclab_arena.utils.pose import Pose
-    from isaaclab_arena_curobo.curobo_planner_utils import make_curobo_planner, sync_object_poses_in_robot_base_frame
+    from isaaclab_arena_curobo.utils.planner_utils import make_curobo_planner, sync_object_poses_in_robot_base_frame
 
     env, embodiment = _build_droid_two_object_env(
         args_cli,
@@ -203,12 +203,12 @@ def _run_sync_pose_check(args_cli) -> bool:
 
 def _run_reachability_check(args_cli) -> bool:
     """Place one object in-reach and one out-of-reach; return True iff IK agrees with that split."""
-    from isaaclab_arena_curobo.curobo_planner_utils import (
+    from isaaclab_arena_curobo.utils.ik_solver_utils import solve_ik_feasibility
+    from isaaclab_arena_curobo.utils.planner_utils import (
         make_curobo_planner,
         sync_object_poses_in_robot_base_frame,
         top_down_grasp_pose_from_env,
     )
-    from isaaclab_arena_curobo.ik_solver_utils import solve_ik_feasibility
 
     env, embodiment = _build_droid_two_object_env(args_cli)
     planner = make_curobo_planner(env, embodiment, env_id=0)
